// job-placement-platform/apps/frontend/components/companies/types.ts
export interface Company {
  id: string;
  name: string;
  email: string;
  website?: string;
  logo?: string;
  postlimit: number;
  companytypeId: string;
  userId: string;
  incorporationLink?: string | null;
  blacklisted: boolean;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
}